Catkeys: unicode in headers and fingerprint
This patch updates the catkeys format with support for unicode in
headers. The parser would break if there were unicode characters in for
example the language name.

Furthermore, the format now has support for fingerprint calculation.
This means that if the tools change the contents of the file (like
remove or change strings), the header will now be correct and the file
will be parseable on Haiku.